fixed refactoring
