+++ basics/src/test/java/basics/ampeg/AmpegClassicSeriesSvt410HlfProductPage.java
@@ -2,9 +2,7 @@ package basics.ampeg;
 
 import org.openqa.selenium.By;
 import org.openqa.selenium.WebDriver;
-import org.openqa.selenium.WebElement;
-
-import basics.pageObjects.PageObject;
+import basics.framework.PageObject;
 
 public class AmpegClassicSeriesSvt410HlfProductPage extends PageObject {
 
@@ -20,5 +18,4 @@ public class AmpegClassicSeriesSvt410HlfProductPage extends PageObject {
 		
 		return handlingWeight;
 	}
-
 }
