+++ basics/src/test/java/basics/ampeg/AmpegSvtProSeriesProductPage.java
@@ -4,7 +4,8 @@ import org.openqa.selenium.By;
 import org.openqa.selenium.WebDriver;
 import org.openqa.selenium.WebElement;
 
-import basics.pageObjects.PageObject;
+import basics.framework.PageObject;
+
 
 public class AmpegSvtProSeriesProductPage extends PageObject {
